Refactored to use live data
