+++ frontend/Components/UserPage.js
@@ -1,10 +1,9 @@
 import { Text, View, Image, FlatList } from "react-native";
 import { StatusBar } from "expo-status-bar";
 import { TouchableOpacity } from "react-native-gesture-handler";
-import { db, auth } from "../test/testdataindex";
-import { useEffect } from "react";
-import { doc, addDoc,getDoc, collection } from "firebase/firestore";
-import { changeUser, fetchUser, fetchUsers } from "../models/model_mock";
+import { changeUser, fetchUser, fetchUsers } from "../models/model_users";
+import styles from "../styles/UserPageStyles";
+import { useState } from "react";
 
 export default function UserPage({
   navigation,
@@ -12,9 +11,14 @@ export default function UserPage({
     params: { username },
   },
 }) {
-  const id = "KVoXaLUNZ511ihffCTVB"
-  const userData = fetchUser(id)
-  const comments = (Object.values(userData[id].reviews))
+  const [comments , setComments] = useState([])
+  const id = "2gHWLG7WkAZgK8iQjvEUjCmGvYG3"
+  fetchUser(id)
+  .then((userData)=>{
+    const newComments = (!Object.values(userData[id].reviews)) ? [] : Object.values(userData[id].reviews)
+    setComments(newComments)
+  })
+  .catch((err)=>console.log(err))
 
   return (
     <View style={styles.pageContainer}>
@@ -22,7 +26,7 @@ export default function UserPage({
         <View style={styles.imageContainer}>
           <Image
             style={styles.image}
-            source={require("../assets/shrek.webp")}
+            source={require('../assets/shrek.webp')}
           ></Image>
           <Text style={styles.username}>{username}</Text>
         </View>
@@ -55,7 +59,7 @@ export default function UserPage({
       <TouchableOpacity
         style={styles.button}
         onPress={() => {
-          navigation.navigate("Message", { username });
+          navigation.navigate('Message', { username });
         }}
       >
         <Text style={styles.text}>Message</Text>
@@ -80,91 +84,3 @@ export default function UserPage({
     </View>
   );
 }
-
-import { StyleSheet } from "react-native";
-import { useState } from "react";
-
-const styles = StyleSheet.create({
-  pageContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#694fad",
-    width: "100%",
-  },
-  image: {
-    height: 150,
-    width: 150,
-    backgroundColor: "white",
-    borderRadius: 100,
-  },
-  imageContainer: {
-    flexDirection: "column",
-    alignItems: "center",
-    padding: 10,
-  },
-  username: {
-    margin: 15,
-    fontWeight: "bold",
-    color: "white",
-    fontSize: 18,
-  },
-  header: {
-    flex: 1,
-
-    width: "100%",
-  },
-  accountInfo: {
-    width: "100%",
-    flex: 1,
-    backgroundColor: "#F1EEED",
-    borderTopLeftRadius: 50,
-    borderTopRightRadius: 50,
-  },
-  accountInfoHeader: {
-    textAlign: "center",
-    margin: 15,
-    color: "#694fad",
-    fontWeight: "bold",
-    fontSize: 22,
-    padding: 10,
-  },
-  accountInfoGrid: {
-    flexDirection: "column",
-  },
-  accountInfoGridTop: {
-    flexDirection: "row",
-    justifyContent: "flex-start",
-    paddingLeft: 5,
-  },
-
-  accountInfoHeadings: {
-    color: "grey",
-    fontWeight: "bold",
-    fontSize: 18,
-  },
-  comments: {
-    flex: 1,
-    backgroundColor: "#F1EEED",
-    width: "100%",
-    alignContent: "center",
-  },
-  signOutContainer: { backgroundColor: "white" },
-  signOutText: {},
-  button: {
-    backgroundColor: "#694fad",
-    borderRadius: 10,
-    fontWeight: "700",
-    fontSize: 16,
-    justifyContent: "center",
-    marginHorizontal: 5,
-    height: 50,
-    padding: 10,
-  },
-  text: {
-    textAlign: "center",
-    color: "#F0EDF6",
-    fontWeight: "500",
-    fontSize: 16,
-  },
-});
